Added column Total Sales on Sales Page

+++ Models/OrdersModel.cs
@@ -9,7 +9,7 @@ namespace SoftEngWebEmployee.Models
     {
         public int Order_ID { get; set; }
         public int CustomerID { get; set; }
-        public double OrderTotalPrice { get; set; }
+        public int OrderTotalPrice { get; set; }
         public string OrderStatus { get; set; }
         public string OrderDate { get; set; }
         public int TotalNumberOfOrders { get; set; }
